feat(mariadb): implementation for MariaDB Reusing existing classes
