+++ lib/dialects/abstract/query-generator.js
@@ -1056,7 +1056,7 @@ class QueryGenerator {
         return `(${quotedColumn}->>'${pathStr}')`;
 
       case 'mariadb':
-        pathStr = `${['$'].concat(paths).join('.')}`;
+        pathStr = ['$'].concat(paths).join('.');
         return `json_unquote(json_extract(${quotedColumn},'${pathStr}'))`;
 
       case 'sqlite':
