+++ lib/dialects/mariadb/query.js
@@ -149,10 +149,10 @@ class Query extends AbstractQuery {
           //ONLY TRUE IF @auto_increment_increment is set to 1 !!
           //Doesn't work with GALERA => each node will reserve increment (x for first server, x+1 for next node ...
           const startId = data[this.getInsertIdField()];
-          result = [];
+          result = new Array(data.affectedRows);
           const pkField = this.model.rawAttributes[this.model.primaryKeyAttribute].field;
-          for (let i = startId; i < startId + data.affectedRows; i++) {
-            result.push({[pkField]: i});
+          for (let i = 0; i < data.affectedRows; i++) {
+            result[i] = {[pkField]: startId + i};
           }
           return [result, data.affectedRows];
         }
@@ -186,9 +186,8 @@ class Query extends AbstractQuery {
       result = {};
 
       for (const _result of data) {
-        const enumRegex = /^enum/i;
         result[_result.Field] = {
-          type: enumRegex.test(_result.Type) ? _result.Type.replace(enumRegex,
+          type: _result.Type.toLowerCase().startsWith('enum') ? _result.Type.replace(/^enum/i,
             'ENUM') : _result.Type.toUpperCase(),
           allowNull: _result.Null === 'YES',
           defaultValue: _result.Default,
@@ -355,5 +354,3 @@ class Query extends AbstractQuery {
 }
 
 module.exports = Query;
-module.exports.Query = Query;
-module.exports.default = Query;
