+++ lib/dialects/mysql/query-generator.js
@@ -6,6 +6,25 @@ const AbstractQueryGenerator = require('../abstract/query-generator');
 const util = require('util');
 const Op = require('../../operators');
 
+
+const jsonFunctionRegex = /^\s*((?:[a-z]+_){0,2}jsonb?(?:_[a-z]+){0,2})\([^)]*\)/i;
+const jsonOperatorRegex = /^\s*(->>?|@>|<@|\?[|&]?|\|{2}|#-)/i;
+const tokenCaptureRegex = /^\s*((?:([`"'])(?:(?!\2).|\2{2})*\2)|[\w\d\s]+|[().,;+-])/i;
+const foreignKeyFields = 'CONSTRAINT_NAME as constraint_name,\n'
+  + '  CONSTRAINT_NAME as constraintName,\n'
+  + '  CONSTRAINT_SCHEMA as constraintSchema,\n'
+  + '  CONSTRAINT_SCHEMA as constraintCatalog,\n'
+  + '  TABLE_NAME as tableName,\n'
+  + '  TABLE_SCHEMA as tableSchema,\n'
+  + '  TABLE_SCHEMA as tableCatalog,\n'
+  + '  COLUMN_NAME as columnName,\n'
+  + '  REFERENCED_TABLE_SCHEMA as referencedTableSchema,\n'
+  + '  REFERENCED_TABLE_SCHEMA as referencedTableCatalog,\n'
+  + '  REFERENCED_TABLE_NAME as referencedTableName,\n'
+  + '  REFERENCED_COLUMN_NAME as referencedColumnName ';
+
+const typeWithoutDefault = new Set(['BLOB', 'TEXT', 'GEOMETRY', 'JSON']);
+
 class MySQLQueryGenerator extends AbstractQueryGenerator {
   constructor(options) {
     super(options);
@@ -57,29 +76,28 @@ class MySQLQueryGenerator extends AbstractQueryGenerator {
     const attrStr = [];
 
     for (const attr in attributes) {
-      if (attributes.hasOwnProperty(attr)) {
-        const dataType = attributes[attr];
-        let match;
+      if (!attributes.hasOwnProperty(attr)) continue;
+      const dataType = attributes[attr];
+      let match;
 
-        if (dataType.includes('PRIMARY KEY')) {
-          primaryKeys.push(attr);
+      if (dataType.includes('PRIMARY KEY')) {
+        primaryKeys.push(attr);
 
-          if (dataType.includes('REFERENCES')) {
-            // MySQL doesn't support inline REFERENCES declarations: move to the end
-            match = dataType.match(/^(.+) (REFERENCES.*)$/);
-            attrStr.push(`${this.quoteIdentifier(attr)} ${match[1].replace('PRIMARY KEY', '')}`);
-            foreignKeys[attr] = match[2];
-          } else {
-            attrStr.push(`${this.quoteIdentifier(attr)} ${dataType.replace('PRIMARY KEY', '')}`);
-          }
-        } else if (dataType.includes('REFERENCES')) {
+        if (dataType.includes('REFERENCES')) {
           // MySQL doesn't support inline REFERENCES declarations: move to the end
           match = dataType.match(/^(.+) (REFERENCES.*)$/);
-          attrStr.push(`${this.quoteIdentifier(attr)} ${match[1]}`);
+          attrStr.push(`${this.quoteIdentifier(attr)} ${match[1].replace('PRIMARY KEY', '')}`);
           foreignKeys[attr] = match[2];
         } else {
-          attrStr.push(`${this.quoteIdentifier(attr)} ${dataType}`);
+          attrStr.push(`${this.quoteIdentifier(attr)} ${dataType.replace('PRIMARY KEY', '')}`);
         }
+      } else if (dataType.includes('REFERENCES')) {
+        // MySQL doesn't support inline REFERENCES declarations: move to the end
+        match = dataType.match(/^(.+) (REFERENCES.*)$/);
+        attrStr.push(`${this.quoteIdentifier(attr)} ${match[1]}`);
+        foreignKeys[attr] = match[2];
+      } else {
+        attrStr.push(`${this.quoteIdentifier(attr)} ${dataType}`);
       }
     }
 
@@ -339,7 +357,7 @@ class MySQLQueryGenerator extends AbstractQueryGenerator {
       };
     }
 
-    const attributeString = attribute.type.toString({ escape: this.escape.bind(this) });
+    const attributeString = attribute.type.toString({escape: this.escape.bind(this)});
     let template = attributeString;
 
     if (attribute.allowNull === false) {
@@ -351,7 +369,9 @@ class MySQLQueryGenerator extends AbstractQueryGenerator {
     }
 
     // BLOB/TEXT/GEOMETRY/JSON cannot have a default value
-    if (!['BLOB', 'TEXT', 'GEOMETRY', 'JSON'].includes(attributeString) && attribute.type._binary !== true && Utils.defaultValueSchemable(attribute.defaultValue)) {
+    if (!typeWithoutDefault.has(attributeString)
+      && attribute.type._binary !== true
+      && Utils.defaultValueSchemable(attribute.defaultValue)) {
       template += ` DEFAULT ${this.escape(attribute.defaultValue)}`;
     }
 
@@ -427,10 +447,6 @@ class MySQLQueryGenerator extends AbstractQueryGenerator {
       return false;
     }
 
-    const jsonFunctionRegex = /^\s*((?:[a-z]+_){0,2}jsonb?(?:_[a-z]+){0,2})\([^)]*\)/i;
-    const jsonOperatorRegex = /^\s*(->>?|@>|<@|\?[|&]?|\|{2}|#-)/i;
-    const tokenCaptureRegex = /^\s*((?:([`"'])(?:(?!\2).|\2{2})*\2)|[\w\d\s]+|[().,;+-])/i;
-
     let currentIndex = 0;
     let openingBrackets = 0;
     let closingBrackets = 0;
@@ -472,35 +488,12 @@ class MySQLQueryGenerator extends AbstractQueryGenerator {
     }
 
     // Check invalid json statement
-    hasInvalidToken |= openingBrackets !== closingBrackets;
-    if (hasJsonFunction && hasInvalidToken) {
+    if (hasJsonFunction && (hasInvalidToken || openingBrackets !== closingBrackets)) {
       throw new Error(`Invalid json statement: ${stmt}`);
     }
 
     // return true if the statement has valid json function
     return hasJsonFunction;
-  }
-
-  /**
-   *  Generates fields for getForeignKeysQuery
-   * @returns {string} fields
-   * @private
-   */
-  _getForeignKeysQueryFields() {
-    return [
-      'CONSTRAINT_NAME as constraint_name',
-      'CONSTRAINT_NAME as constraintName',
-      'CONSTRAINT_SCHEMA as constraintSchema',
-      'CONSTRAINT_SCHEMA as constraintCatalog',
-      'TABLE_NAME as tableName',
-      'TABLE_SCHEMA as tableSchema',
-      'TABLE_SCHEMA as tableCatalog',
-      'COLUMN_NAME as columnName',
-      'REFERENCED_TABLE_SCHEMA as referencedTableSchema',
-      'REFERENCED_TABLE_SCHEMA as referencedTableCatalog',
-      'REFERENCED_TABLE_NAME as referencedTableName',
-      'REFERENCED_COLUMN_NAME as referencedColumnName'
-    ].join(',');
   }
 
   /**
@@ -511,9 +504,9 @@ class MySQLQueryGenerator extends AbstractQueryGenerator {
    * @returns {string}            The generated sql query.
    * @private
    */
-  getForeignKeysQuery(tableName, schemaName) {
-    return `SELECT ${this._getForeignKeysQueryFields()} FROM INFORMATION_SCHEMA.KEY_COLUMN_USAGE where TABLE_NAME = '${tableName  /* jshint ignore: line */
-    }' AND CONSTRAINT_NAME!='PRIMARY' AND CONSTRAINT_SCHEMA='${schemaName}' AND REFERENCED_TABLE_NAME IS NOT NULL;`; /* jshint ignore: line */
+  getForeignKeysQuery(table, schemaName) {
+    const tableName = table.tableName || table;
+    return `SELECT ${foreignKeyFields} FROM INFORMATION_SCHEMA.KEY_COLUMN_USAGE where TABLE_NAME = '${tableName}' AND CONSTRAINT_NAME!='PRIMARY' AND CONSTRAINT_SCHEMA='${schemaName}' AND REFERENCED_TABLE_NAME IS NOT NULL;`;
   }
 
   /**
@@ -525,18 +518,15 @@ class MySQLQueryGenerator extends AbstractQueryGenerator {
    * @private
    */
   getForeignKeyQuery(table, columnName) {
-    const tableName = table.tableName || table;
-    const schemaName = table.schema;
+    const quotedSchemaName = table.schema ? wrapSingleQuote(table.schema) : '';
+    const quotedTableName = wrapSingleQuote(table.tableName || table);
+    const quotedColumnName = wrapSingleQuote(columnName);
 
-    return `SELECT ${this._getForeignKeysQueryFields()
-    } FROM INFORMATION_SCHEMA.KEY_COLUMN_USAGE`
-      + ` WHERE (REFERENCED_TABLE_NAME = ${wrapSingleQuote(tableName)
-      }${schemaName ? ` AND REFERENCED_TABLE_SCHEMA = ${wrapSingleQuote(schemaName)}`: ''
-      } AND REFERENCED_COLUMN_NAME = ${wrapSingleQuote(columnName)
-      }) OR (TABLE_NAME = ${wrapSingleQuote(tableName)
-      }${schemaName ? ` AND TABLE_SCHEMA = ${wrapSingleQuote(schemaName)}`: ''
-      } AND COLUMN_NAME = ${wrapSingleQuote(columnName)
-      } AND REFERENCED_TABLE_NAME IS NOT NULL)`;
+    return `SELECT ${foreignKeyFields} FROM INFORMATION_SCHEMA.KEY_COLUMN_USAGE`
+      + ` WHERE (REFERENCED_TABLE_NAME = ${quotedTableName}${table.schema
+        ? ` AND REFERENCED_TABLE_SCHEMA = ${quotedSchemaName}`
+        : ''} AND REFERENCED_COLUMN_NAME = ${quotedColumnName})
+       OR (TABLE_NAME = ${quotedTableName}${table.schema ? ` AND TABLE_SCHEMA = ${quotedSchemaName}` : ''} AND COLUMN_NAME = ${quotedColumnName} AND REFERENCED_TABLE_NAME IS NOT NULL)`;
   }
 
   /**
@@ -548,9 +538,10 @@ class MySQLQueryGenerator extends AbstractQueryGenerator {
    * @private
    */
   dropForeignKeyQuery(tableName, foreignKey) {
-    return `ALTER TABLE ${this.quoteTable(tableName)} DROP FOREIGN KEY ${this.quoteIdentifier(foreignKey)};`;
+    return `ALTER TABLE ${this.quoteTable(tableName)} 
+      DROP FOREIGN KEY ${this.quoteIdentifier(foreignKey)};`;
   }
-};
+}
 
 // private methods
 function wrapSingleQuote(identifier) {
