+++ lib/dialects/mysql/query-interface.js
@@ -12,13 +12,13 @@ const Promise = require('../../promise');
 const sequelizeErrors = require('../../errors');
 
 /**
-  A wrapper that fixes MySQL's inability to cleanly remove columns from existing tables if they have a foreign key constraint.
+ A wrapper that fixes MySQL's inability to cleanly remove columns from existing tables if they have a foreign key constraint.
 
-  @param  {string} tableName     The name of the table.
-  @param  {string} columnName    The name of the attribute that we want to remove.
-  @param  {Object} options
+ @param  {string} tableName     The name of the table.
+ @param  {string} columnName    The name of the attribute that we want to remove.
+ @param  {Object} options
 
-  @private
+ @private
  */
 function removeColumn(tableName, columnName, options) {
   options = options || {};
@@ -28,7 +28,7 @@ function removeColumn(tableName, columnName, options) {
       tableName,
       schema: this.sequelize.config.database
     }, columnName),
-    Object.assign({ raw: true }, options)
+    Object.assign({raw: true}, options)
   )
     .then(([results]) => {
       //Exclude primary key constraint
@@ -38,38 +38,40 @@ function removeColumn(tableName, columnName, options) {
       }
       return Promise.map(results, constraint => this.sequelize.query(
         this.QueryGenerator.dropForeignKeyQuery(tableName, constraint.constraint_name),
-        Object.assign({ raw: true }, options)
+        Object.assign({raw: true}, options)
       ));
     })
     .then(() => this.sequelize.query(
       this.QueryGenerator.removeColumnQuery(tableName, columnName),
-      Object.assign({ raw: true }, options)
+      Object.assign({raw: true}, options)
     ));
 }
 
-
 function removeConstraint(tableName, constraintName, options) {
-  const sql = this.QueryGenerator.showConstraintsQuery(tableName.tableName ? tableName : {
-    tableName,
-    schema: this.sequelize.config.database
-  }, constraintName);
+  const sql = this.QueryGenerator.showConstraintsQuery(
+    tableName.tableName ? tableName : {
+      tableName,
+      schema: this.sequelize.config.database
+    }, constraintName);
 
-  return this.sequelize.query(sql, Object.assign({}, options, { type: this.sequelize.QueryTypes.SHOWCONSTRAINTS }))
+  return this.sequelize.query(sql, Object.assign({}, options,
+    {type: this.sequelize.QueryTypes.SHOWCONSTRAINTS}))
     .then(constraints => {
       const constraint = constraints[0];
       let query;
-      if (constraint && constraint.constraintType) {
-        if (constraint.constraintType === 'FOREIGN KEY') {
-          query = this.QueryGenerator.dropForeignKeyQuery(tableName, constraintName);
-        } else {
-          query = this.QueryGenerator.removeIndexQuery(constraint.tableName, constraint.constraintName);
-        }
+      if (!constraint || !constraint.constraintType) {
+        throw new sequelizeErrors.UnknownConstraintError(
+          {
+            message: `Constraint ${constraintName} on table ${tableName} does not exist`,
+            constraint: constraintName,
+            table: tableName
+          });
+      }
+
+      if (constraint.constraintType === 'FOREIGN KEY') {
+        query = this.QueryGenerator.dropForeignKeyQuery(tableName, constraintName);
       } else {
-        throw new sequelizeErrors.UnknownConstraintError({
-          message: `Constraint ${constraintName} on table ${tableName} does not exist`,
-          constraint: constraintName,
-          table: tableName
-        });
+        query = this.QueryGenerator.removeIndexQuery(constraint.tableName, constraint.constraintName);
       }
 
       return this.sequelize.query(query, options);
