+++ lib/query-interface.js
@@ -7,7 +7,6 @@ const DataTypes = require('./data-types');
 const SQLiteQueryInterface = require('./dialects/sqlite/query-interface');
 const MSSSQLQueryInterface = require('./dialects/mssql/query-interface');
 const MySQLQueryInterface = require('./dialects/mysql/query-interface');
-const MariaDBQueryInterface = require('./dialects/mariadb/query-interface');
 const PostgresQueryInterface = require('./dialects/postgres/query-interface');
 const Transaction = require('./transaction');
 const Promise = require('./promise');
@@ -527,11 +526,9 @@ class QueryInterface {
         // mssql needs special treatment as it cannot drop a column with a default or foreign key constraint
         return MSSSQLQueryInterface.removeColumn.call(this, tableName, attributeName, options);
       case 'mysql':
-        // mysql needs special treatment as it cannot drop a column with a foreign key constraint
-        return MySQLQueryInterface.removeColumn.call(this, tableName, attributeName, options);
       case 'mariadb':
-        // drop a column with a foreign key constraint
-        return MariaDBQueryInterface.removeColumn.call(this, tableName, attributeName, options);
+        // mysql/mariadb need special treatment as it cannot drop a column with a foreign key constraint
+        return MySQLQueryInterface.removeColumn.call(this, tableName, attributeName, options);
       default:
         return this.sequelize.query(this.QueryGenerator.removeColumnQuery(tableName, attributeName), options);
     }
@@ -854,10 +851,9 @@ class QueryInterface {
 
     switch (this.sequelize.options.dialect) {
       case 'mysql':
-        //Mysql does not support DROP CONSTRAINT. Instead DROP PRIMARY, FOREIGN KEY, INDEX should be used
-        return MySQLQueryInterface.removeConstraint.call(this, tableName, constraintName, options);
       case 'mariadb':
-        return MariaDBQueryInterface.removeConstraint.call(this, tableName, constraintName, options);
+        //does not support DROP CONSTRAINT. Instead DROP PRIMARY, FOREIGN KEY, INDEX should be used
+        return MySQLQueryInterface.removeConstraint.call(this, tableName, constraintName, options);
       case 'sqlite':
         return SQLiteQueryInterface.removeConstraint.call(this, tableName, constraintName, options);
       default:
