+++ test/integration/dialects/mariadb/associations.test.js
@@ -58,8 +58,8 @@ describe('[MariaDB Specific] Associations', () => {
       this.users = null;
       this.tasks = null;
 
-      this.User.belongsToMany(this.Task, {as: 'Tasks',through: 'UserTasks'});
-      this.Task.belongsToMany(this.User, {as: 'Users',through: 'UserTasks'});
+      this.User.belongsToMany(this.Task, {as: 'Tasks', through: 'UserTasks'});
+      this.Task.belongsToMany(this.User, {as: 'Users', through: 'UserTasks'});
 
       const self = this,
         users = [],
@@ -79,78 +79,5 @@ describe('[MariaDB Specific] Associations', () => {
 
     });
 
-    describe('addDAO / getModel', () => {
-      beforeEach(function() {
-        const self = this;
-
-        self.user = null;
-        self.task = null;
-
-        return self.User.findAll()
-          .then(_users => {
-            return self.Task.findAll().then(_tasks => {
-              self.user = _users[0];
-              self.task = _tasks[0];
-            });
-          });
-      });
-
-      it('should correctly add an association to the dao', function() {
-        const self = this;
-
-        return self.user.getTasks().then(_tasks => {
-          expect(_tasks.length).to.equal(0);
-          return self.user.addTask(self.task).then(() => {
-            return self.user.getTasks().then(_tasks => {
-              expect(_tasks.length).to.equal(1);
-            });
-          });
-        });
-      });
-    });
-
-    describe('removeDAO', () => {
-      beforeEach(function() {
-        const self = this;
-
-        self.user = null;
-        self.tasks = null;
-
-        return self.User.findAll().then(_users => {
-          return self.Task.findAll().then(_tasks => {
-            self.user = _users[0];
-            self.tasks = _tasks;
-          });
-        });
-      });
-
-      it('should correctly remove associated objects', function() {
-        const self = this;
-
-        return self.user.getTasks()
-          .then(__tasks => {
-            expect(__tasks.length).to.equal(0);
-            return self.user.setTasks(self.tasks);
-          })
-          .then(() => {
-            return self.user.getTasks();
-          })
-          .then(_tasks => {
-            expect(_tasks.length).to.equal(self.tasks.length);
-            return self.user.removeTask(self.tasks[0]);
-          })
-          .then(() => {
-            return self.user.getTasks();
-          })
-          .then(_tasks => {
-            expect(_tasks.length).to.equal(self.tasks.length - 1);
-            return self.user.removeTasks([self.tasks[1], self.tasks[2]]);
-          })
-          .then(() => self.user.getTasks())
-          .then(_tasks => {
-            expect(_tasks).to.have.length(self.tasks.length - 3);
-          });
-      });
-    });
   });
 });
